Add PriceCatalog service, Order service and API Gateway

+++ FreakyFashionServices.Basket/Models/Domain/CartContent.cs
@@ -10,7 +10,7 @@ namespace FreakyFashionServices.Basket.Models.Domain
         public int Id { get; set; }
         public string Name { get; set; }
         public string Description { get; set; }
-        public decimal UnitPrice { get; set; }
+        public int UnitPrice { get; set; }
         public int Quantity { get; set; }
     }
 }
